+++ packages/semi-d2c-typings/src/core.ts
@@ -77,3 +77,31 @@ export interface SandboxParams {
     | 'svelte'
     | 'vue-cli';
 }
+
+export interface D2CInspect {
+  /** 提示名，如 semiComponent  */
+  inspectName: string;
+  /** 中文提示名 */
+  inspectChineseName?: string;
+  /** 提示名的解释，用于解释这个类型提示的作用  */
+  inspectDescription?: string;
+  /** 提示类型，如 error  */
+  inspectType: 'info' | 'warning' | 'error';
+  /** 提示标题，可传 node.name 或 node.name 拼接的字符串 */
+  inspectTitle: string;
+  /** 提示的详细内容，如不知道写什么，可以与 inspectDescription 一致 */
+  inspectMessage: string;
+  /**
+   * 被检查的节点 id 们，id 不能为空
+   * 可传单个 id，也可以传多个 id，多个 id 会被作为一次检查结果
+   */
+  nodeIds: string[];
+  /** 检查文档  */
+  docURL?: string;
+  /** 示意图 */
+  imgURL?: string;
+  /** 是否默认展示 inspect 线框  */
+  defaultVisible?: boolean;
+  /** inspect 线框颜色 */
+  color?: string;
+}