+++ packages/create-semi-d2c-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @douyinfe/create-semi-d2c-plugin
 
+## 0.0.4
+
+### Patch Changes
+
+- feat: add inspectDraft and modifyInspects hooks, update ts config file
+
 ## 0.0.3
 
 ### Patch Changes